+++ lib/input/reader/sync_batcher.go
@@ -141,7 +141,13 @@ func (p *SyncBatcher) CloseAsync() {
 // WaitForClose blocks until either the reader is finished closing or a timeout
 // occurs.
 func (p *SyncBatcher) WaitForClose(tout time.Duration) error {
-	return p.r.WaitForClose(tout)
+	tstop := time.Now().Add(tout)
+	err := p.r.WaitForClose(time.Until(tstop))
+	p.batcher.CloseAsync()
+	if err != nil {
+		return err
+	}
+	return p.batcher.WaitForClose(time.Until(tstop))
 }
 
 //------------------------------------------------------------------------------